+++ src/Yakka/DefaultCalculator.cs
@@ -22,35 +22,30 @@ namespace Yakka
     /// <summary>
     /// Contains methods to calculate the working hours and the break.
     /// </summary>
-    public static class WorkingHoursCalculator
+    public class DefaultCalculator : ICalculator
     {
-        /// <summary>
-        /// Calculates the working hours using the specified <see cref="IWorkingHoursCalculator"/> and the given
-        /// start of the work day.
-        /// </summary>
-        /// <param name="configuration">The working hours configuration.</param>
-        /// <param name="currentTime">The current time.</param>
-        /// <returns>An instance of <see cref="WorkingHoursCalculation"/> containing the information about the working hours.</returns>
-        /// <exception cref="ArgumentNullException"><c>calculator</c> is <c>null</c>.</exception>
-        /// <exception cref="InvalidOperationException">The working hours could not be calculated.</exception>
-        public static WorkingHoursCalculation Calculate(WorkingHoursConfiguration configuration, DateTime currentTime)
+        /// <inheritdoc />
+        public WorkingHoursCalculation Calculate(WorkingHoursConfiguration configuration, DateTime currentTime)
         {
             if (configuration == null)
             {
                 throw new ArgumentNullException(nameof(configuration));
             }
 
-            var result = new WorkingHoursCalculation();
+            var result = new WorkingHoursCalculation(configuration);
             result.Configuration = configuration;
 
             if (configuration.BreakMode == BreakMode.Automatic)
             {
-                (result.CalculatedWorkingHours, result.CalculatedBreak) = configuration.WorkingHoursCalculator.Calculate(configuration.StartTime, currentTime);
+                (result.CalculatedWorkingHours, result.CalculatedBreak, result.Warning) = configuration.WorkingHoursCalculator.Calculate(configuration.StartTime, currentTime);
             }
             else if (configuration.BreakMode == BreakMode.Manual)
             {
+                (_, _, string? warning) = configuration.WorkingHoursCalculator.Calculate(configuration.StartTime, currentTime);
+
                 result.CalculatedBreak = configuration.ManualBreakTime;
                 result.CalculatedWorkingHours = currentTime - configuration.StartTime - result.CalculatedBreak;
+                result.Warning = warning;
             }
             else
             {
@@ -65,7 +60,7 @@ namespace Yakka
                 while (workedTimeSpan.TotalHours < hoursWorked)
                 {
                     specificEndOfWorkDay = specificEndOfWorkDay.AddMinutes(1);
-                    (workedTimeSpan, _) = configuration.WorkingHoursCalculator.Calculate(configuration.StartTime, specificEndOfWorkDay);
+                    (workedTimeSpan, _, _) = configuration.WorkingHoursCalculator.Calculate(configuration.StartTime, specificEndOfWorkDay);
                 }
 
                 result.FullHoursWorked.Add(hoursWorked, specificEndOfWorkDay);