+++ tests/Yakka.Tests/Calculation/GermanStatic45WorkingHoursCalculatorTestscs.cs
@@ -27,6 +27,18 @@ namespace Yakka.Tests.Calculation
     public class GermanStatic45WorkingHoursCalculatorTestscs
     {
         /// <summary>
+        /// Tests the required metadata.
+        /// </summary>
+        [TestMethod]
+        public void MetadataTest()
+        {
+            var calulator = new GermanStatic45WorkingHoursCalculator();
+
+            Assert.IsFalse(string.IsNullOrWhiteSpace(calulator.Title));
+            Assert.IsFalse(string.IsNullOrWhiteSpace(calulator.Description));
+        }
+
+        /// <summary>
         /// Checks the calculation.
         /// </summary>
         [TestMethod]
@@ -36,7 +48,7 @@ namespace Yakka.Tests.Calculation
 
             DateTime start;
             DateTime end;
-            (TimeSpan workTimeSpan, TimeSpan breakTimeSpan) result;
+            (TimeSpan workTimeSpan, TimeSpan breakTimeSpan, string? warning) result;
 
             // 6h, 0m
             end = DateTime.Now;